+++ src/main/java/com/flowiee/pms/repository/sales/OrderDetailRepository.java
@@ -10,9 +10,9 @@ import com.flowiee.pms.entity.sales.OrderDetail;
 import java.util.List;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
     @Query("from OrderDetail d where d.order.id=:orderId")
-    List<OrderDetail> findByOrderId(@Param("orderId") Integer orderId);
+    List<OrderDetail> findByOrderId(@Param("orderId") Long orderId);
 
     @Query("from OrderDetail d where d.order.id = :orderId and d.productDetail.id = :productVariantId")
     OrderDetail findByOrderIdAndProductVariantId(@Param("orderId") int orderId, @Param("productVariantId") int productVariantId);
